Playground stages unify into one API-surface column (D429)
Every stage is now the same shape: a centered column on the hero card's
gutter — input, Run, the result of that run — with common parameters
inline and the rest behind a closed Advanced panel. The text stage drops
the chat (ChatStage -> TextStage, one prompt / one reply, prompt in the
URL), the image stage drops the gallery strip, and the right settings
rail dies across all four.
