+++ DECISIONS.md
@@ -1083,3 +1083,4 @@ Security layer (token, Origin/Host validation, sandboxed bridge — see threat n
 | # | Decision | Choice | Rationale / rejected alternatives |
 |---|---|---|---|
 | D428 | **The Playground sidebar's model rows become spacious CARDS carrying exactly four things — nickname, full name, size, and a Download CTA — grouped under one collapsible `<details open>` section per capability whose header wears a per-capability icon (chat bubble / picture / mic / magnifier, `playground/capabilityIcons.tsx`)** | Reverses two earlier calls by design request. The old row ("a row, not a card: LM Studio style") packed the curator's note, a fit verdict, a state dot and a status word into a 264px rail; the redesign trades that density for legibility: the note and the fit sentence still exist where the decision is actually made (the stage header and its state line), so the card drops them without losing them. **The second line never repeats the first**: `nickname` falls back to `label` for cached entries, so the full-name line renders only when it differs, else the repo id, else nothing. **The card is a div-as-button, not a `<button>`** — the Download CTA lives inside it and a button inside a button is markup browsers are free to mangle; keyboard selection is Enter/Space on the card. The CTA reuses the stage's own `downloadAiModel` + `refreshAiRuntime` path (and its ErrorBanner — a card has no room for a server sentence), disables to "Downloading…" off `runtime.downloading`, and a downloaded model shows a quiet "Downloaded" instead — the existing catalog shrink-detection refetches when the pull lands, no new plumbing. Clicking Download also SELECTS the model, deliberately: the stage is where the progress bar lives. **Sections are native `<details open>`** — free collapse state, correct semantics, no React state to persist — with the default marker replaced by a right-hand caret so the capability icon owns the left edge. The unavailable-group reason (HF-8), the empty-shortlist sentence and the "Your downloads" caption all survive inside the details body — collapse must not become the second way a ruled-out capability goes invisible, which is why the reason line lives BELOW the summary: a collapsed group still opens to its explanation. Icons are the tab's own file, not `groups.ts`, because Home imports that module eagerly and JSX would drag markup into the front-door bundle; not `MenuIcons`, because these four glyphs name TASKS, not actions, and belong to the one surface that speaks task words. Sidebar widens 264→300px to pay for the card padding. **Rejected: nesting the CTA in the row-button** (invalid HTML). **Rejected: hiding undownloaded state behind hover** — the CTA is the one action the card exists to offer. **Rejected: React-state collapse** with persistence; nobody asked for remembered collapse, and `<details>` is deletable the day someone does. |
+| D429 | **Every Playground stage becomes ONE shape — a centered column on the hero card's own gutter (`min(680px, 100% - 64px)`) reading as an API surface: input, Run, the result of THAT run, with the common parameters inline under the composer and everything else behind a closed-by-default Advanced `<details>`. The text stage drops the chat (ChatStage → TextStage: one prompt, one reply, no history sent), the image stage drops the gallery strip (one image, the current run's), and the right settings rail + its `Controls` toggle are deleted from all four** | The four stages had four layouts (chat column + rail, image column + rail, transcribe column + rail, embed column alone), and the tab's actual reader is someone testing what one `fused.ai` call returns — a transcript and a thumbnail strip are app furniture answering a question nobody here asked, while the rail put temperature behind a sidebar on wide screens and a toggle on narrow ones. One column, same gutter as the hero, means the header and its stage read as one card stack on every capability. **The prompt now rides the URL** (`prompt`, the image stage's precedent): a one-off prompt is SETUP, not transcript, so the chat rule ("the URL carries the setup, never the transcript") reclassifies it rather than being broken by it. History is sent empty; `splitThink`, per-reply stats, Copy, Stop, the 409 load dance and cancel-on-unmount all survive verbatim — this moved a surface, not the wire. Common/advanced split per stage: text = temperature + max tokens up front (system prompt, top-p behind); image = aspect + speed chips (sizes, steps, guidance, seed behind — the seed-reuse caption button stays, it was the strip's only function that mattered); transcribe = task (language, VAD, word timestamps behind); embed has no knobs and just takes the column. `.pg-stage` becomes the scroller — a one-shot column has no inner log, so the hero scrolls with content. **Rejected: keeping chat behind a toggle** — two modes is two surfaces to keep honest for a tab whose exit ramp ("Build an app") already leads to anything stateful. **Rejected: a shared StageShell component** — the shape is three class names; a wrapper would abstract less code than it adds. **Rejected: keeping the rail on wide screens only** — the point is one shape everywhere, and a layout that changes at 1100px is two shapes wearing one name. |
